cloud/server/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
cryptography==41.0.4
